+++ RELEASE_NOTES.md
@@ -5,6 +5,14 @@
 - The language, federation and date format settings have moved from the browser to the main application window (5.1.0)
 - Plugins are now installed and uninstalled from the main application window (5.1.0)
 
+## Tournaments
+
+- Exclusion of players (or teams) who have not played 50% of the games (or matches) in accordance with FIDE 6.6 in round robins (5.1.0)
+
+## Pairings
+
+- Ask if a round should be marked as completed for individual round robins (5.1.0)
+
 ## Documents
 
 - Place cards can now be edited with an embedded editor (5.1.0)